requirements.txt
numpy==1.26.4
librosa==0.10.2.post1
scikit-learn==1.5.2
matplotlib==3.8.0
keras==3.5.0
tensorflow==2.17.1
soundfile==0.12.1
streamlit==1.40.1
openai-whisper==20240930
